feat : shopping basket fix
