+++ src/pages/_app.tsx
@@ -4,12 +4,7 @@ import Head from 'next/head';
 import globalStyles from 'styles/global-styles';
 import {ChakraProvider} from '@chakra-ui/react';
 import {RecoilRoot} from 'recoil';
-import {firestore} from '../firebase';
-import {useEffect} from 'react';
 function App({Component, pageProps}: AppProps) {
-  useEffect(() => {
-    console.log(firestore);
-  });
   return (
     <RecoilRoot>
       <ChakraProvider>
